pyproject.toml
[tool.poetry]
name = "bluescraper"
version = "0.1.0"
description = "Bluescraper is a Python web scraping tool that simplifies content extraction using configuration files, enhancing efficiency and reusablity."
authors = ["Malte Sauerwein <m.sauerwein@live.de>"]
readme = "README.md"
packages = [{include = "bluescraper", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
black = "^24.2.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
mypy-extensions = "^1.0.0"
pylint = "^3.0.3"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
types-beautifulsoup4 = "^4.12.0.20240106"
types-requests = "^2.31.0.20240125"
pyyaml = "^6.0.1"
pydantic = "^2.6.1"
pre-commit = "^3.6.1"
types-pyyaml = "^6.0.12.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
